+++ main.tf
@@ -18,19 +18,18 @@
   Provider configuration
  *****************************************/
 provider "google" {
-  credentials = "${file(var.credentials_file_path)}"
 }
 
 locals {
-  //  secret_project = "${var.project_name}"
-  shared_bucket = "shared-${var.env}-secrets"
-  app_bucket    = "${var.application_name}-${var.env}-secrets"
-  bucket_name   = "${var.shared == "true" ? local.shared_bucket : local.app_bucket}"
-  object_path   = "${var.secret}.txt"
+  shared_bucket = "shared-${var.project_id}-${var.env}-secrets"
+  app_bucket    = substr("${var.project_id}-${var.application_name}-${var.env}-secrets", 0, 64)
+  bucket_name   = var.shared ? local.shared_bucket : local.app_bucket
+  object_path   = length(regexall(".+[.](txt|json)$", var.secret)) > 0 ? var.secret : "${var.secret}.txt"
 }
 
 module "secret" {
   source = "./modules/gcs-object"
-  bucket = "${local.bucket_name}"
-  path   = "${local.object_path}"
+  bucket = local.bucket_name
+  path   = local.object_path
+
 }