+++ modules/create-secret/outputs.tf
@@ -10,7 +10,15 @@ output "self_link" {
   value       = google_storage_bucket_object.secret.self_link
 }
 
-output "output_name" {
+output "secret_name" {
   description = "The name of the object. Use this field in interpolations with"
   value       = google_storage_bucket_object.secret.output_name
+}
+
+output "test1" {
+  value = local.file_type
+}
+
+output "test2" {
+  value = local.content_type
 }