+++ test/setup/main.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 Google LLC
+ * Copyright 2018 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,19 +18,16 @@ module "project" {
   source  = "terraform-google-modules/project-factory/google"
   version = "~> 3.0"
 
-  name              = "ci-datalab"
+  name              = "ci-secrets"
   random_project_id = "true"
   org_id            = var.org_id
   folder_id         = var.folder_id
   billing_account   = var.billing_account
 
   activate_apis = [
-    "cloudresourcemanager.googleapis.com",
     "storage-api.googleapis.com",
-    "serviceusage.googleapis.com",
+    "cloudresourcemanager.googleapis.com",
     "compute.googleapis.com",
-    "iap.googleapis.com",
-    "iam.googleapis.com",
-    "iamcredentials.googleapis.com",
+    "serviceusage.googleapis.com"
   ]
 }