+++ test/setup/outputs.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 Google LLC
+ * Copyright 2018 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,9 +20,5 @@ output "project_id" {
 
 output "sa_key" {
   value     = google_service_account_key.int_test.private_key
-  sensitive = true
-}
-
-output "datalab_service_account_email" {
-  value = module.project.service_account_email
+//  sensitive = true
 }