+++ test/setup/versions.tf
@@ -15,13 +15,9 @@
  */
 
 terraform {
-  required_version = ">= 0.12"
+  required_version = "~> 0.12.6"
 }
 
 provider "google" {
-  version = "~> 2.13.0"
-}
-
-provider "google-beta" {
   version = "~> 2.13.0"
 }