+++ CHANGELOG.md
@@ -8,3 +8,6 @@ project adheres to [Semantic Versioning](http://semver.org/).
 ## [v0.1.0]
 ### Added
 - Initial module release
+
+- added project name to bucket name (original name for bucket)
+- fixed bug with incorrect name for shared bucket